move eternity-related files to their own folder
