+++ src/core/eternity/eternity-challenge.js
@@ -1,6 +1,6 @@
-import { DC } from "./constants";
+import { DC } from "../constants";
 import { deepmergeAll } from "@/utility/deepmerge";
-import { GameMechanicState } from "./game-mechanics";
+import { GameMechanicState } from "../game-mechanics";
 
 export function startEternityChallenge() {
   initializeChallengeCompletions();
