+++ src/core/eternity/index.js
@@ -1,0 +1,4 @@
+export * from "./eternity.js";
+export * from "./eternity-challenge.js";
+export * from "./time-theorems.js";
+export * from "./dilation.js";
