+++ src/core/globals.js
@@ -67,15 +67,12 @@ export * from "./time-studies";
 export * from "./dimboost";
 export * from "./sacrifice";
 export * from "./infinity/index.js";
+export * from "./eternity/index.js";
 export * from "./normal-challenges";
-export * from "./eternity";
-export * from "./eternity-challenge";
 export * from "./reality";
-export * from "./time-theorems";
 export * from "./reality-upgrades";
 export * from "./imaginary-upgrades";
 export * from "./perks";
-export * from "./dilation";
 export * from "./black-hole";
 export * from "./machines";
 export * from "./devtools";
